requirements.txt
aioredis==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
cffi==1.16.0
channels==3.0.4
channels-redis==3.3.1
constantly==23.10.4
cryptography==42.0.5
daphne==3.0.2
Django==4.1.13
django-environ==0.11.1
djongo==1.3.6
dnspython==2.6.1
gunicorn==19.1.1
hiredis==2.3.2
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
msgpack==1.0.8
mysqlclient==2.2.4
paho-mqtt==1.6.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pymongo==3.12.3
pyOpenSSL==24.1.0
pytz==2024.1
service-identity==24.1.0
six==1.16.0
sqlparse==0.2.4
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.10.0
websockets==12.0
zope.interface==6.2
